feat: fetcher interface with CPC impelementation
A fetcher is responsible for gathering source data from a provider and
making it available to an extractor whose job will be to turn it into
zarr_jsons. Fetchers use fsspec so that they can be deployed in a
distributed environment.

Updates on this branch have been squashed to a single commit. With any
luck this new commit will be signed with my ssh key.