feat: allow extractor to return multiple single zarr jsons
It came up during the presentation with Arbol, that they need the
ability to return multiple zarr jsons from a single source file. Since
allowying for that involves modifying the contract provided by the
Extractor interface, it made sense to go ahead and make that change now,
since the more things that get built on top of them, the harder it is to
change a core interface specification.